app/requirements.txt
django-chartjs
django-extensions
django-qr-code
django-tables2
djangorestframework
gunicorn
pyprowl
python-dotenv
validators
